requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.6.2.post1
asyncer==0.0.7
bidict==0.23.1
cachetools==5.5.0
certifi==2024.8.30
chainlit==1.3.2
charset-normalizer==3.4.0
chevron==0.14.0
click==8.1.7
dataclasses-json==0.6.7
Deprecated==1.2.15
docstring_parser==0.16
environs==9.5.0
fastapi==0.115.5
filelock==3.16.1
filetype==1.2.0
fsspec==2024.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.13.1
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
Lazify==0.4.0
literalai==0.0.623
marshmallow==3.23.1
milvus-lite==2.4.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
opentelemetry-api==1.28.1
opentelemetry-exporter-otlp==1.28.1
opentelemetry-exporter-otlp-proto-common==1.28.1
opentelemetry-exporter-otlp-proto-grpc==1.28.1
opentelemetry-exporter-otlp-proto-http==1.28.1
opentelemetry-instrumentation==0.49b1
opentelemetry-proto==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
packaging==23.2
pandas==2.2.3
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.0
pymilvus==2.4.9
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-multipart==0.0.9
python-rapidjson==1.20
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.4.5
shapely==2.0.6
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
syncer==2.0.3
tokenizers==0.20.3
tomli==2.1.0
tqdm==4.67.0
transformers==4.46.2
tritonclient==2.51.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
uptrace==1.27.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.25.0
vertexai==1.71.1
watchfiles==0.20.0
wrapt==1.16.0
wsproto==1.2.0
zipp==3.21.0
